feat(reader): the note composer docks beside the document
Taking a note opened a dialog over the passage it was about, and a modal
can only ever hold one capture: the next selection had nowhere to go.

NoteComposer is the composer; variant="sheet" is what a page with no room
beside it opens, variant="docked" is the reader's third panel face. A
second capture appends into the open draft. The panel's Notes tab lists
this document's notes when nothing is being captured.

make verify-dock, 33 checks green. Fired: rendering the composer as a
sheet again gives "1 dialogs", "0 composers", and the second selection
cannot be made at all -- the sheet is over the transcript. Dropping
captureKey gives "3 -> 3 quoted lines".

Do not unmount the panel to collapse it; hiding it is what keeps a
streaming answer and an unsaved draft alive.
